Change when configxml opens
Show it before MainWindow
